requirements.txt
dearpygui
hanlp
pandas
regex
opencc
requests
hanzidentifier
pyperclip
pywinstyles
google-generativeai
numpy>=1.25.0,<2.0.0